rename collate to modules
